Add conversion from list to matrix for weights when single component
A single component model has weights stored as a data frame with a single column.  sapply simplifies these down to a list of numeric vectors instead of the expected matrix of models x components.  This if block checks for the list output and converts it to the expected matrix in the correct orientation and correct row name.