Enhancement for #111
enhancement(DA): provided a more informative error message when a Y vector containing NAs is passed to any DA method

Added a check within `Check.entry.pls()` and `Check.entry.wrapper.mint.block()` which checks if discriminant analysis is being done and if unmapped Y dataframe has any rows which sum to 0 (no associated class). If so, stops with specific error message.

Added `DA = TRUE` to `internal_wrapper.mint()` calls in `block.splsda()`, `mint.plsda()` and `mint.splsda()` so that the new checks will be engaged.

Removed the `BPPARAM` parameter from the `repeat_cv()` call in `MCVfold.block.splsda()` as this was raising an error in the tests. Functionality is seemingly unaffected