PR addressing Issue #246 (#247)
refactor: adjusted `mean_centering_per_study()` to force it to return global `center` and `scale` attributes as well as the per study attributes

fix: added the `weights` and `indY` objects to the returned value of `mint.block.splsda()`. These are required as part of `predict()` which is used in `auroc()`

fix: added the `weights` and `indY` objects to the returned value of `mint.block.plsda()`. These are required as part of `predict()` which is used in `auroc()`

test: added tests for `mint.block.plsda()` and `mint.block.splsda()` to maintain package coverage